Moved the ProtoDescriptorStrings class out of `grpc_server.cc` and into `proto_descriptor_strings.cc`
