+++ src/grpc_server.cc
@@ -9,6 +9,7 @@
 #include <future>
 #include <grpcpp/impl/server_initializer.h>
 #include "lv_proto_server_reflection_plugin.h"
+#include "proto_descriptor_strings.h"
 
 //---------------------------------------------------------------------
 //---------------------------------------------------------------------
@@ -271,43 +272,5 @@ namespace grpc_labview
             _server = nullptr;
         }
         grpc_labview::ProtoDescriptorStrings::getInstance()->deleteInstance();
-    }
-
-    //---------------------------------------------------------------------
-    //---------------------------------------------------------------------
-
-    // Initialize the static members
-    ProtoDescriptorStrings* ProtoDescriptorStrings::m_instance = nullptr;
-    std::mutex ProtoDescriptorStrings::m_mutex;
-
-    // Return the static class instance. Thread safe.
-    ProtoDescriptorStrings* ProtoDescriptorStrings::getInstance() {
-        std::unique_lock<std::mutex> lock(m_mutex);
-        if (m_instance == nullptr) {
-            m_instance = new ProtoDescriptorStrings();
-        }
-        return m_instance;
-    }
-
-    // Get the descriptor string
-    std::vector<std::string> ProtoDescriptorStrings::getAllDescriptors() {
-        std::unique_lock<std::mutex> lock(m_mutex);
-        return m_descriptors;
-    }
-
-    // Set the descriptor string
-    void ProtoDescriptorStrings::addDescriptor(std::string str) {
-        std::unique_lock<std::mutex> lock(m_mutex);
-        m_refcount++;
-        m_descriptors.push_back(str);
-    }
-
-    // Delete the instance based on the refcount
-    void ProtoDescriptorStrings::deleteInstance() {
-        std::unique_lock<std::mutex> lock(m_mutex);
-        if (!--m_refcount) {
-            delete m_instance;
-            m_instance = nullptr;
-        }
     }
 }
